requirements.txt
aiohttp-jinja2~=1.6
Jinja2~=3.1.5
aiohttp~=3.11.11
PyYAML~=6.0.2
SQLAlchemy~=2.0.37
alembic~=1.14.1
asyncpg~=0.30.0
aiohttp_session~=2.12.1
cryptography~=44.0.0
python-dotenv~=1.0.1
